feat(workflows): add auto-update meta information for songs (main)
The commit includes the addition of an automated workflow to update the meta information for songs. The Git commands used are 'git config user.name github-actions' and 'git config user.email github-actions@github.com'. The changes made in this commit are related to the file '.github/workflows/songs_auto_update_meta.yml'. The commit message is '[Bot] I have added all of the meta information about the content to it.'